stl support init
